commit refactoring editorial code
