+++ apps/nextjs-website/src/components/react-components/src/index-props.ts
@@ -1,4 +1,4 @@
-import { HeroProps } from './components/Hero/Hero.types';
-import { EditorialProps } from './components/Editorial/Editorial';
+import { HeroProps } from './utils/Components.types';
+import { EditorialProps } from './utils/Components.types';
 
 export type { HeroProps, EditorialProps };
